+++ attributes/default.rb
@@ -7,6 +7,10 @@ default['opsline-rails-app']['unicorn']['timeout'] = '45'
 default['opsline-rails-app']['unicorn']['worker_processes'] = '2'
 
 default['opsline-rails-app']['ruby']['provider'] = 'rbenv'
-default['opsline-rails-app']['ruby']['version'] = '2.1.2'
-default['opsline-rails-app']['ruby']['gems'] = []
-
+default['opsline-rails-app']['ruby']['versions'] = ['2.1.2']
+default['opsline-rails-app']['ruby']['gems'] = {
+  'bundler' => {
+    'version' => nil,
+    'ruby_versions' => ['2.1.2']
+  }
+}