+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.0
 # Adding worker container
+* Ability to install multiple versions of ruby via rbenv
+* Support for rvm
 
 ## 0.1.0:
 * Initial release